Back-end: Implemented new module - sockets server - first phase Back-end: Made socket server port configurable and environment dependent Back-end: Integrated the new socket server into our web application
This commit is the first phase of the new server module - sockets server. The purpose of the socket server is to deal with real-time messages sent between the front-end and the back-end.
The commit contains: the module itself (built on top of Socket.IO library), dependencies requirements in the package.json, configurable listen port for the socket server and the integration with our web application.

#294, #302, #303