Back-end: Separated the functionality of the sockets server from the sockets server file Back-end: Implemented: Sockets-server: Has to provide functionality and not an enum Back-end: Added changing notification read state functionality to the notification module
The commit contains a new module - socketsServerHandlers.js which includes the functionalities provided by the socket server so the socket server listen to messages and triggers the relevant function in the socketsServerHandlers. The separation was done for 2 purposes: the rest server can now use the function as well (avoiding code duplication) and for testing purposes (unit tests).
In addition, in order to improve the abstraction, the sockets server doesn't expose the types enum but provide functions to the clients.
Furthermore, the notification module is now responsible for changing the read state of a notification instead of manual editing the property in the client side.

#350, #357,  #359