Back-end: Added private messaging module
The module includes the schema and the basic functionality that depends on the schema of a private message.
The private message (AKA message) object is the core unit of the private messaging feature. a message object represents a message sent from user X and it is a part of a conversation (so it's a weak entity).

#368