Back-end: Updated modules following the #418 changes Testing: Back-end: Created tests for PATCH /users/notification/:id Back-end: Added Support multiple ids on PATCH /users/notifications/:id
Changed the id list passed to route PATCH /users/notifications/:id to be query instead of a single param so the new route is PATCH /users/notifications.

Furthermore, Modified user model to support input of array for marking notifications as read. In addition, updated all places that used the relevant function to send an array.

This commit includes testing of the updated route.

#418, #426, #444