Back-end: Added new route to fetch all reviews (without aggregation)
Added a new route to get all reviews near-by the given coordinates without aggregation. In addition, modified the current tests. The new route is part of integrating the back-end and the front-end implementation of the reviews feature as described in issue: Back-end: Reviews routes should return the format expected by the front-end

#513, #516