Front-end: part of integrating reviews page with backend Front-end: Fixed Bug: Reviews pros and cons are not displayed in front end
This commit prepares the Review Tile component for integrating with the back-end i.e. the publisher properties(name, image...) will be passed from the parent component that communicates  with the back-end. In addition, minor bug fix - the review object contains Pros and Cons keys and not pros and cons (case sensitive).

#501, #527